Move enc/dec scalar fallbacks straight into asm and split on build tags instead of at runtime.
This consolidates the encoding tests, gets rid of the initialization logic -- and the fallbacks ended up being faster, on top of there being room for an actual optimization pass. However, amd64 builds require SSE2 as a result. It is available on virtually every amd64 platform, but should this assumption not hold true in the wild, sno will panic during startup. Yet another fallback may need to be provided if that ends up being the case.

Signed-off-by: Michał Chojnacki <m.chojnacki@muyo.io>