Base time calculations on wall clock only.
Simplifies the logic a bit (gets rid of tracking the mono clock offset etc.) and we really don't need the monotonic OS clock as long as we guarantee one stored sample of the wall clock to only ever increase monotonically.

Regressions and restoration during a regression are now covered by deterministic tests. As this needed the clock to be injectable into a generator anyways, opening up the possibility to provide custom time sources to Generators (custom time units/epoch and/or cluster-wide synchronized clocks etc.) is now an option for the future (although utility methods on ID, e.g. Time() and Timestamp() would need a different approach, since they run on static assumptions about the epoch and units).

Snapshot.Now is retained as a timestamp of the Snapshot (in epoch units) to make comparisons possible, but is otherwise not used anymore by the library itself.

Signed-off-by: Michał Chojnacki <m.chojnacki@muyo.io>