Allow compiler optimization of Generator.applyTimestamp().
PutUint32 and PutUint64 in encoding/binary are compiler assisted in their signature and boil down to a wide load + shift + bswap. While in our use case the uint64 for this is suboptimal, 3 instructions is still much simpler than the scalar 16 needed otherwise. Arguably this is the more idiomatic approach as well.

Microbenchmarking note @go1.14: Actual synthetic difference in benchmarks is within noise (-0.015ns) when applyTimestamp() is tested on its own. However, when run in the apply* chain of functions (alone) it shaved off 1ns (3.8ns down to 2.8ns), whereas benchmarked in the full context of Generator.New() it resulted in a reduction of 0.7ns (13.2ns down to 12.5ns). Actual benchmark results in the README may get an update along with the bench suite separated out into a package.