Remove internal errInvalidDataSize and errPartitionPoolExhausted pointers.
It's unreasonable to have them allocated from the get go, wasting memory. errPartitionPoolExhausted is an extreme edge case whereas errInvalidDataSize is unlikely and in most cases denotes a logic error (lack of input validation or straight out programming error) instead of an unpredictable runtime fault condition.

As a byproduct, InvalidDataSizeError now contains a public Size field, with the size/length of the value which triggerered the error.