Optimize time fetching (initial pass).
- Remove the time source from generators. It was introduced to make tests easier, allowing to swap it out for fake time. For an intermediary solution the time source now gets monkey patched during tests instead (see note below about planned changes, though). This reduces the size of the Generator structs by a function pointer (32/64bits) and allows us to avoid a dynamic dispatch during ID generation (which shaves off 0.8ns);
- windows/amd64: Now includes an optimized, hand written assembly function to get the OS time, bypassing Go's runtime now() which is written to work with time.Now() instead and has different requirements. This shaves off about a dozen or so instructions - including a fairly expensive back-and-forth conversion between time units - and coupled with previous change to time sources brings generation time on windows/amd64 down to 8.8nsec/op (down from 11.8ns/op before this commit and 13.2ns at v1.0.0).

Note: The monkey patching currently introduces an external dependency which I'll happily get rid of due to not being able to explicitly separate out test deps from productions deps with go modules. This will likely be in favour of simply relying on build tags to swap out the underlying time source for tests. However, this blocked by a pending test refactoring pass and potentially splitting some code out into an internal package. Additionally, the patching code doesn't play well with inlining which is why tests need to be run with -gcflags=all=-l (to disable inlining) while this code is in (which I'll just as happily get rid of to keep the test as close to production code as possible).