Remove internal bookkeeping data pointers
This slighly reduces the size of Generators, simplifies the initialization logic and pushes less work on the GC. The changed addressing appeared to have no performance impact, leading me to question why I stored them as pointers initially.

This also drops the regression branch mutex pointer along with its initialization logic. In practice Generators should've been 'noCopy' from the start, but this was undocumented until now. 'go vet' will also pick up on the mutex and report when a Generator gets copied.

Generators still need to be initialized at this stage. While a set of commits switching the seqOverflow logic to be lazily initiated (albeit slower) including some changes to the synchronization flow may make the zero value of a Generator usable in the sense that it won't panic, being scoped to a zero Partition this would have arguable utility as we definitely don't want to make the partition mutable after initialization.