Simplify time code structure
This roughly follows what the std lib does with its faketime build tag. Since the wrapper in the !test builds is so thin, it becomes a compiled away zero-cost abstraction while still allowing us to easily swap it out with a dynamic dispatch in tests. This allows us to get rid of the duplicated implementations.