Provide temporary fallback for Snotime via time.Now() on linux/amd64/go1.17 due to Go's assembly implementation being removed on this target.
This is a (hopefully temporary) performance regression on this target: the call involves getting the monotonic clock time and back-and-forth conversions while we only need the wall clock time, which is what runtime.walltime1 used to provide. However, the removal from the runtime package necessitates backporting a significant portion of Go's runtime assembly for this target, where the previous bypass to achieve the perf boost over time.Now() was trivial in comparison, so I decided against maintaining such backport and will look into a more manageable solution instead.